requirements.txt
flask
requests
pyyaml
watchdog
psutil
schedule
pywinrm
uvicorn[standard] fastapi